third_party/WebKit/LayoutTests/fast/scroll-behavior/scroll-customization/scrollstate-basic.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>ScrollState constructor behaves correctly</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

test(function() {
  assert_true('ScrollState' in window, "'ScrollState' in window");
}, "These tests only work with scroll customization enabled.");

if ('ScrollState' in window) {
  test(function() {
    var scrollState = new ScrollState();
    assert_equals(scrollState.deltaX, 0);
    assert_equals(scrollState.deltaY, 0);
    assert_equals(scrollState.deltaGranularity, 0);
    assert_equals(scrollState.velocityX, 0);
    assert_equals(scrollState.velocityY, 0);
    assert_equals(scrollState.inInertialPhase, false);
    assert_equals(scrollState.isBeginning, false);
    assert_equals(scrollState.isEnding, false);
    assert_equals(scrollState.fromUserInput, false);
  }, "Empty constructor behaves correctly.");

  test(function() {
    var init = {
      deltaX: 12.5,
      deltaY: 14.9,
      positionX: 16.82,
      positionY: 82.17,
      velocityX: 23.7,
      velocityY: 2.5,
      isBeginning: true,
      isInInertialPhase: true,
      isEnding: true,
      fromUserInput: true,
      isDirectManipulation: true,
      deltaGranularity: 148.3
    };
    var scrollState = new ScrollState(init);
    assert_equals(scrollState.deltaX, init.deltaX);
    assert_equals(scrollState.deltaY, init.deltaY);
    assert_equals(scrollState.positionX, Math.floor(init.positionX));
    assert_equals(scrollState.positionY, Math.floor(init.positionY));
    assert_equals(scrollState.velocityX, init.velocityX);
    assert_equals(scrollState.velocityY, init.velocityY);
    assert_equals(scrollState.isBeginning, init.isBeginning);
    assert_equals(scrollState.inInertialPhase, init.isInInertialPhase);
    assert_equals(scrollState.isEnding, init.isEnding);
    assert_equals(scrollState.fromUserInput, init.fromUserInput);
    assert_equals(scrollState.isDirectManipulation, init.isDirectManipulation);
    assert_equals(scrollState.deltaGranularity, init.deltaGranularity);
  }, "Constructor behaves correctly.");

  test(function() {
    var scrollState = new ScrollState();
    scrollState.isBeginning = true;
    assert_equals(scrollState.isBeginning, false);
    scrollState.fromUserInput = true;
    assert_equals(scrollState.fromUserInput, false);
  }, "attributes are read only");
}
</script>
</body>
</html>
